ymgportal/requirements.txt
asgiref==3.9.1
boto3==1.41.2
botocore==1.41.2
Django==5.2.5
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.15.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
